requirements.txt
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.3.0
tqdm==4.64.1
wandb==0.13.5
nibabel>=3.2.0
SimpleITK>=2.1.0
scipy>=1.9.0
